move enum out of utils directory
