+++ projects/swimlane/ngx-ui/src/lib/enums/keyboard-keys.enum.ts
@@ -1,0 +1,10 @@
+export enum KeyboardKeys {
+  ENTER = 'Enter',
+  ESCAPE = 'Escape',
+  DELETE = 'Delete',
+  BACKSPACE = 'Backspace',
+  ARROW_DOWN = 'ArrowDown',
+  ARROW_UP = 'ArrowUp',
+  ARROW_LEFT = 'ArrowLeft',
+  ARROW_RIGHT = 'ArrowRight'
+}
