+++ projects/swimlane/ngx-ui/src/lib/utils/index.ts
@@ -2,4 +2,3 @@ export * from './debounce';
 export * from './throttle';
 export * from './id';
 export * from './position';
-export * from './keys';
