--- projects/swimlane/ngx-ui/src/lib/utils/keys.ts
@@ -1,10 +1,0 @@
-export enum KeyboardKeys {
-  ENTER = 'Enter' as any,
-  ESCAPE = 'Escape' as any,
-  DELETE = 'Delete' as any,
-  BACKSPACE = 'Backspace' as any,
-  ARROW_DOWN = 'ArrowDown' as any,
-  ARROW_UP = 'ArrowUp' as any,
-  ARROW_LEFT = 'ArrowLeft' as any,
-  ARROW_RIGHT = 'ArrowRight' as any
-}
